+++ ChangeLog.md
@@ -2,5 +2,5 @@
 0.1
 ===
 
-- Harry.Fixed orphans 
+- Harpie.Fixed orphans